fix LeaveOrgModal
