+++ src/universal/components/BillingLeaderActionMenu.tsx
@@ -1,17 +1,17 @@
-import React from 'react'
-import MenuWithShortcuts from 'universal/components/MenuWithShortcuts'
-import MenuItemWithShortcuts from 'universal/components/MenuItemWithShortcuts'
-import {BILLING_LEADER} from 'universal/utils/constants'
-import SetOrgUserRoleMutation from 'universal/mutations/SetOrgUserRoleMutation'
-import LeaveOrgModal from 'universal/modules/userDashboard/components/LeaveOrgModal/LeaveOrgModal'
-import RemoveFromOrgModal from 'universal/modules/userDashboard/components/RemoveFromOrgModal/RemoveFromOrgModal'
+import {BillingLeaderActionMenu_organization} from '__generated__/BillingLeaderActionMenu_organization.graphql'
+import {BillingLeaderActionMenu_organizationUser} from '__generated__/BillingLeaderActionMenu_organizationUser.graphql'
+import React, {lazy} from 'react'
 import {createFragmentContainer, graphql} from 'react-relay'
+import MenuItemWithShortcuts from 'universal/components/MenuItemWithShortcuts'
+import MenuWithShortcuts from 'universal/components/MenuWithShortcuts'
 import withAtmosphere, {
   WithAtmosphereProps
 } from 'universal/decorators/withAtmosphere/withAtmosphere'
+import RemoveFromOrgModal from 'universal/modules/userDashboard/components/RemoveFromOrgModal/RemoveFromOrgModal'
+import SetOrgUserRoleMutation from 'universal/mutations/SetOrgUserRoleMutation'
+import {BILLING_LEADER} from 'universal/utils/constants'
 import withMutationProps, {WithMutationProps} from 'universal/utils/relay/withMutationProps'
-import {BillingLeaderActionMenu_organizationUser} from '__generated__/BillingLeaderActionMenu_organizationUser.graphql'
-import {BillingLeaderActionMenu_organization} from '__generated__/BillingLeaderActionMenu_organization.graphql'
+import LoadableModal from './LoadableModal'
 
 interface Props extends WithMutationProps, WithAtmosphereProps {
   closePortal: () => void
@@ -19,6 +19,10 @@ interface Props extends WithMutationProps, WithAtmosphereProps {
   organizationUser: BillingLeaderActionMenu_organizationUser
   organization: BillingLeaderActionMenu_organization
 }
+
+const LeaveOrgModal = lazy(() =>
+  import(/* webpackChunkName: 'LeaveOrgModal' */ 'universal/modules/userDashboard/components/LeaveOrgModal/LeaveOrgModal')
+)
 
 const BillingLeaderActionMenu = (props: Props) => {
   const {
@@ -59,9 +63,9 @@ const BillingLeaderActionMenu = (props: Props) => {
       )}
       {viewerId === userId &&
         !isViewerLastBillingLeader && (
-          <LeaveOrgModal
-            orgId={orgId}
-            userId={userId}
+          <LoadableModal
+            LoadableComponent={LeaveOrgModal}
+            queryVars={{orgId}}
             toggle={<MenuItemWithShortcuts label='Leave Organization' />}
           />
         )}
