+++ src/universal/modules/notifications/components/RequestNewUser/RequestNewUser.js
@@ -2,7 +2,7 @@ import styled, {css} from 'react-emotion'
 import PropTypes from 'prop-types'
 import React from 'react'
 import {createFragmentContainer} from 'react-relay'
-import withRouter from 'react-router-dom/es/withRouter'
+import {withRouter} from 'react-router-dom'
 import defaultStyles from 'universal/modules/notifications/helpers/styles'
 import ApproveToOrgMutation from 'universal/mutations/ApproveToOrgMutation'
 import ui from 'universal/styles/ui'
