+++ src/universal/modules/notifications/components/TaskInvolves.js
@@ -2,7 +2,7 @@ import {convertFromRaw, Editor, EditorState} from 'draft-js'
 import PropTypes from 'prop-types'
 import React, {Component} from 'react'
 import {createFragmentContainer} from 'react-relay'
-import withRouter from 'react-router-dom/es/withRouter'
+import {withRouter} from 'react-router-dom'
 import OutcomeCardStatusIndicator from 'universal/modules/outcomeCard/components/OutcomeCardStatusIndicator/OutcomeCardStatusIndicator'
 import editorDecorators from 'universal/components/TaskEditor/decorators'
 import ClearNotificationMutation from 'universal/mutations/ClearNotificationMutation'
