+++ src/universal/modules/userDashboard/components/OrganizationRow/OrganizationRow.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import defaultOrgAvatar from 'universal/styles/theme/images/avatar-organization.svg'
 import {PERSONAL, PRO, PRO_LABEL} from 'universal/utils/constants'
-import withRouter from 'react-router-dom/es/withRouter'
+import {withRouter} from 'react-router-dom'
 import plural from 'universal/utils/plural'
 import styled from 'react-emotion'
 import ui from 'universal/styles/ui'
